NamedArgs.pm: Allow to work with 5.20 and above
Left braces "{" that are used literally will eventually need to be
escaped, most simply by preceding them with a backslash.  In the
meantime, their use raises a deprecation warning.

Right braces don't need this, but no harm is done by adding backslashes,
and the resulting symmetry may prevent confusion, so this commit adds
those as well.  Feel free to remove them if that is your preference.